style: Improve Makefile
This improves the Makefile tol run the application in a container for
reproducability. However, we still use an actual binary as the input for
whether or not we need to create a new container.
